Modularize task tests
This probably isn't the best way to structure the tests, it's just
a first step on the road to a better splitting of this.
